Move `Median` and `StdDev` out of `Crypto` namespace
