+++ WalletWasabi/Crypto/Extensions.cs
@@ -22,25 +22,4 @@ public static class Extensions
 			yield return resultSelector(e1.Current, e2.Current, e3.Current);
 		}
 	}
-
-	public static double Median(this IEnumerable<double> me)
-	{
-		if (!me.Any())
-		{
-			return 0;
-		}
-		var sorted = me.OrderBy(x => x).ToArray();
-		return sorted[sorted.Length / 2];
-	}
-
-	public static double StdDev(this IEnumerable<double> values)
-	{
-		var mean = values.Average();
-		var squaresSum = values
-			.Select(x => x - mean)
-			.Select(x => x * x)
-			.Sum();
-
-		return Math.Sqrt(squaresSum / values.Count());
-	}
 }
